[FEATURE] Make sitePackage editable in Backend Sites module
Add 'fake TCA' config adding the 'sitePackage'
field to the Backend Sites module when editing a
site config to select the attached site
package / extension.

The field is a select drop down that shows all
extensions that start with "site_".

The field can of course still be manually set in
the .yaml file and is not overridden when editing
the site config.