[TASK] Prepare for TYPO3 v13 PageTsConfig event
TYPO3 v12 deprecated the ModifyLoadedPageTsConfigEvent
with https://review.typo3.org/c/Packages/TYPO3.CMS/+/76565
and moved to a similar event in a different namespace.

The patch starts listening for the new event and suppresses
handling the old event in TYPO3 v12 to not do things twice.

This prepares for TYPO3 v13 so code can be easily removed
when extension compatibility TYPO3 v11 is dropped later.